+++ TorpedoJatekClient/Utility/Mesh_OGL3.cpp
@@ -36,7 +36,7 @@ void Mesh::draw()
 {
 	glBindVertexArray(vertexArrayObject);
 
-	glDrawElements(GL_TRIANGLES, indices.size(), GL_UNSIGNED_INT, 0);
+	glDrawElements(GL_TRIANGLES, static_cast<int>(indices.size()), GL_UNSIGNED_INT, 0);
 
 	glBindVertexArray(0);
 }
